+++ layers/Aggregator.py
@@ -1,44 +1,12 @@
+from typing import Union, Callable
+
 import attr
+import attrs
 
-from engine.lexer import Condition
+from engine.run_semgrep import run_semgrep_one
 from layers import Builder
-from layers.dataclass.Attributes import Purity, Visibility
-from layers.dataclass.Components import Variable
-
-
-@attr.s(auto_attribs=True)
-class ResultScope:
-    name: str
-    variable: list[Variable]
-
-
-@attr.s(auto_attribs=True)
-class ContractScope(ResultScope):
-    functions: list[str]
-    libraries: list[str]
-
-
-@attr.s(auto_attribs=True)
-class FunctionScope(ResultScope):
-    parameters: list[Variable]
-    purity: Purity
-    visibility: Visibility
-    payable: bool
-    override: bool
-    modifier: list[str]
-    returns: list[Variable]
-    body: str
-    access_control: list[Condition]
-
-
-@attr.s(auto_attribs=True)
-class FileScope:
-    file_name: str
-    license: str
-    solc_version: str
-    imports: list[str]
-    contract_scope: ContractScope
-    function_scopes: list[FunctionScope]
+from layers.Loader import Loader
+from layers.dataclass.Components import FileScope, ContractScope, DetectionLog, FunctionScope
 
 
 @attr.s(auto_attribs=True)
@@ -104,3 +72,120 @@ class Aggregator:
                 ]
             )
         )
+
+
+@attrs.frozen(auto_attribs=True)
+class ThreatDetectionResult:
+    info: CodeAnalysisAggregation
+    threats: list[DetectionLog]
+
+
+@attrs.define(init=False, auto_attribs=True)
+class ThreatModelBase:
+    loader: Loader
+    target_scope: FileScope | ContractScope | FunctionScope
+    title: str
+    summary: str
+    description: str
+    impact: str  # TODO: enum
+    severity: str  # TODO: enum
+    recommendation: str | None
+    rule: callable
+
+    def __init__(self, target_scope: str | FileScope | ContractScope | FunctionScope):
+        self.override_metadata()
+        self.loader = Loader()
+        if type(target_scope) is str:
+            target_scope = target_scope.lower()
+            patterns = {
+                "file": FileScope,
+                "contract": ContractScope,
+                "function": FunctionScope
+            }
+            if target_scope not in patterns:
+                raise ValueError("Invalid target scope")
+            self.target_scope = patterns[target_scope]
+        # else check if target_scope is a subclass of each of the three classes
+        elif target_scope in [FileScope, ContractScope, FunctionScope]:
+            self.target_scope = target_scope
+        else:
+            raise ValueError("Invalid target scope")
+
+    # rule can be defined as a semgrep rule name or a custom function
+    # semgrep rule runs on the target scope source code
+    # custom function runs on attributes which takes the target scope as an argument
+    def mount_rule(self, rule: Union[str, Callable]):
+        if isinstance(rule, str):
+            if self.target_scope is FunctionScope:
+                self.rule = lambda scope, code: (
+                    [run_semgrep_one(rule, self.loader.cache_content(_scope.body, "sol")) for _scope in scope])
+            else:
+                # TODO: make body for each scope
+                self.rule = lambda scope, code: run_semgrep_one(rule, self.loader.cache_content(code, "sol"))
+        elif callable(rule):
+            # this params forced because the rule function must accept the target scope as an argument
+            self.rule = lambda scope, code: rule(scope, code)
+        else:
+            raise ValueError("Invalid rule type")
+
+    def override_metadata(
+        self,
+        title: str = None,
+        summary: str = None,
+        description: str = None,
+        impact: str = None,
+        severity: str = None,
+        recommendation: str = None
+    ):
+        self.title = title
+        self.summary = summary
+        self.description = description
+        self.impact = impact
+        self.severity = severity
+        self.recommendation = recommendation
+
+    def run(self, code: str) -> DetectionLog:
+        base_log = DetectionLog(
+            title=self.title,
+            summary=self.summary,
+            description=self.description,
+            impact=self.impact,
+            severity=self.severity,
+            recommendation=self.recommendation,
+            scopes=[]
+        )
+        from layers.Aggregator import Aggregator
+
+        res = base_log.scopes
+        _agg = Aggregator().aggregate(code)
+        if self.target_scope is FunctionScope:
+            for _scope in _agg.data.function_scopes:
+                if self.rule(_agg.data.function_scopes, _scope.body):
+                    res.append(_scope)
+        elif self.target_scope is ContractScope:
+            if self.rule(_agg.data.contract_scope, code):
+                res.append(_agg.data.contract_scope)
+        elif self.target_scope is FileScope:
+            if self.rule(_agg.data, code):
+                res.append(_agg.data)
+        else:
+            raise ValueError("Invalid target scope")
+        return base_log
+
+
+def get_analysis_result_with_threats(code: str, rules: list[ThreatModelBase]) -> ThreatDetectionResult:
+    from layers.Aggregator import Aggregator
+    result = Aggregator().aggregate(code)
+    fileScope: FileScope = result.data
+    contractScope: ContractScope = fileScope.contract_scope
+    functionScopes: list[FunctionScope] = fileScope.function_scopes
+
+    if not rules:
+        return ThreatDetectionResult(info=result, threats=[])
+
+    threats = []
+
+    for rule in rules:
+        base_detection_log = rule.run(code)
+        if base_detection_log.scopes:
+            threats.append(base_detection_log)
